fix: first cross chain test

+++ test/unit/RebaseTokenTest.t.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: SEE LICENSE IN LICENSE
 pragma solidity ^0.8.19;
 
-import {Test,console} from "forge-std/Test.sol";
+import {Test, console} from "forge-std/Test.sol";
 import {RebaseToken} from "src/RebaseToken.sol";
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {IAccessControl} from "@openzeppelin/contracts/access/IAccessControl.sol";
@@ -24,7 +24,7 @@ contract RebaseTokenTest is Test {
         vm.stopPrank();
 
         vm.startPrank(MINTER_BURNER);
-        rebase.mint(USER, INITIAL_VALUE,rebase.getInterestRate());
+        rebase.mint(USER, INITIAL_VALUE, rebase.getInterestRate());
         vm.stopPrank();
     }
 
@@ -88,7 +88,7 @@ contract RebaseTokenTest is Test {
 
         // Verify old users keep old rate, new users get new rate
         vm.prank(MINTER_BURNER);
-        rebase.mint(NEW_USER, 1 ether,newRate);
+        rebase.mint(NEW_USER, 1 ether, newRate);
         assertEq(rebase.getUserInterestRate(USER), INITIAL_INTEREST_RATE);
         assertEq(rebase.getUserInterestRate(NEW_USER), newRate);
     }
@@ -111,7 +111,7 @@ contract RebaseTokenTest is Test {
 
     function test_OnlyMinterCanCallMint() external {
         vm.startPrank(address(0x123));
-        uint interestRate = rebase.getUserInterestRate(USER);
+        uint256 interestRate = rebase.getUserInterestRate(USER);
         vm.expectRevert(
             abi.encodeWithSelector(
                 IAccessControl.AccessControlUnauthorizedAccount.selector, address(0x123), rebase.getMintAndBurnRole()
@@ -180,12 +180,13 @@ contract RebaseTokenTest is Test {
         _;
     }
     //TODO: Test this function
+
     function test_MintNewRebaseTokens() external changeTimeStamp {
         uint256 newAmountToBurn = 1 ether;
         vm.startPrank(MINTER_BURNER);
         uint256 userInterestRate = rebase.getUserInterestRate(USER);
         console.log("Fetching interest rate of user:", userInterestRate);
-        rebase.mint(USER, newAmountToBurn,userInterestRate);
+        rebase.mint(USER, newAmountToBurn, userInterestRate);
 
         assertEq(rebase.balanceOf(USER), rebase.principleBalanceOf(USER));
         vm.stopPrank();
@@ -280,7 +281,7 @@ contract RebaseTokenTest is Test {
 
         //New USER -> IR = e4e10, USER -> IR = 5e10
         vm.prank(MINTER_BURNER);
-        rebase.mint(NEW_USER, INITIAL_VALUE,newInterestRate); //Tokens minted using new interest rate
+        rebase.mint(NEW_USER, INITIAL_VALUE, newInterestRate); //Tokens minted using new interest rate
 
         //if USER tries to transfer all the funds to NEW_USER, then USER's IR changes to new Interest rate
         vm.prank(USER);
